+++ R/neon_helper_functions.r
@@ -18,7 +18,7 @@ make_neon_location_table <- function(loc_info, loc_col_names){
   
   # make location table
   table_location <- suppressMessages(
-      ecocomDP::create_location(
+      ecocomDP::create_location_deprecated(
         L0_flat = loc_info,
         location_id = "location_id",
         latitude = "decimalLatitude",