+++ data-raw/create_ants_L1.R
@@ -5,5 +5,8 @@ library(ecocomDP)
 # Read from EDI data repository
 ants_L1 <- read_data(id = "edi.193.4")
 
+# Return as tibble
+ants_L1$edi.193.4$tables <- lapply(ants_L1$edi.193.4$tables, tidyr::as_tibble)
+
 # Save to /data
 usethis::use_data(ants_L1, overwrite = TRUE)